requirements.txt
cognite-sdk-core==2.39.1
